fixed issue that getSocketIndex returned wrong fd, which could not accept second server

+++ include/Server.hpp
@@ -42,7 +42,7 @@ private:
         void setNonBlocking(int fd);
         void cleanup();
         void timeout_clients();
-        std::optional<int> getSocketIndex(int fd) const;
+        std::optional<size_t> getSocketIndex(int fd) const;
 
         void send_cgi_responses(void);
         void epoll_event_loop(int num_events);
